Hamilton_table_2

Standard deviation of cyclical component and correlation with cyclical component of GDP for assorted macroeconomic series.

data("Hamilton_table_2")

Format

A data.frame containing 13 economic time series observations of 5 variables:

**cycle.sd:**

Standard deviation of the cycle component, computed with function yth_filter(x, output = "cycle").

**gdp.cor:**

Correlation of cycle.sd with the cycle.sd of 100 * log(RealGDP).

**random.sd:**

Standard deviation of a random walk, computed by rolling differencing the series by period h, the same h passed to yth_filter(x, output = "cycle") to compute cycle.sd.

**gdp.rand.cor:**

Correlation of random.sd with the random.sd of 100 * log(RealGDP).

Source

Hamilton, James D. (2018). *Why You Should Never Use the Hodrick-Prescott Filter*, pg. 40. http://econweb.ucsd.edu/~jhamilto/hp.pdf

Notes

Filtered series were based on the full sample available for that variable, while correlations were calculated using the subsample of overlapping values for the two indicators. Note that the regression residuals lose the first 11 observations and the random-walk calculations lose the first 8 observations.

Examples

print(Hamilton_table_2)
#>                        cycle.sd gdp.cor random.sd gdp.rand.cor        Sample
#> GDP                        3.38    1.00      3.69         1.00 1947-1/2016-1
#> Consumption                2.85    0.79      3.04         0.82 1947-1/2016-1
#> Investment                13.19    0.84     13.74         0.80 1947-1/2016-1
#> Exports                   10.77    0.33     11.33         0.30 1947-1/2016-1
#> Imports                    9.79    0.77      9.98         0.75 1947-1/2016-1
#> Government-spending        7.13    0.31      8.60         0.38 1947-1/2016-1
#> Employment                 3.09    0.85      3.32         0.85 1947-1/2016-2
#> Unemployment-rate          1.44   -0.81      1.72        -0.79 1948-1/2016-2
#> GDP-Deflator               2.99    0.04      4.11        -0.13 1947-1/2016-1
#> S&P500                    21.80    0.41     22.08         0.38 1950-1/2016-2
#> 10-year-Treasury-yield     1.46   -0.05      1.51         0.08 1953-2/2016-2
#> Fedfunds-rate              2.78    0.33      3.03         0.40 1954-3/2016-2
#> Real-rate                  2.25    0.39      2.60         0.42 1958-1/2014-3